Francois' move (moving all the files)
