+++ client/src/pages/CreateDuel.js
@@ -6,7 +6,7 @@ import Avatar from "../components/Avatar";
 import ButtonFullWidth from "../components/buttons/ButtonFullWidth";
 import Loading from "../components/Loading";
 import PageError from "../components/PageError";
-import { getChallengeableUsers } from "../utils/apiQueries";
+import { getChallengeableUsers } from "../utils/queryUsers";
 
 const CreateDuel = () => {
   const [searchRegex, setSearchRegex] = useState(null);
